refactor: move useValidation to hooks folder
